Старостин Д.Д. №3

+++ StarostinD/Lab0/Lab0.c
@@ -16,17 +16,17 @@ int check(int x1, int y1, int r1, int x2, int y2, int r2) {
 }
 
 int main() {
-	int x1, y1, r1 = 0, x2, y2, r2 = 0;
+	double x1, y1, r1 = 0, x2, y2, r2 = 0;
 	printf("Enter information about the first circle: ");
 	while (r1 <= 0) {
-		scanf_s("%d%d%d", &x1, &y1, &r1);
+		scanf_s("%lf%lf%lf", &x1, &y1, &r1);
 		if (r1 <= 0) {
 			printf("Error! Radius can't be 0 or less. Please enter correct information.\n");
 		}
 	}
 	printf("Enter information about the second circle: ");
 	while (r2 <= 0) {
-		scanf_s("%d%d%d", &x2, &y2, &r2);
+		scanf_s("%lf%lf%lf", &x2, &y2, &r2);
 		if (r2 <= 0) {
 			printf("Error! Radius can't be 0 or less. Please enter correct information.\n");
 		}
